Solve the "song aliases" problem
Easy solution: add all aliases in a separate xml file, and parse it
first. Then we add all songs with aliases to the db. Then when we parse
all songs in sets we know if a title already exists, and don't need to
know if it's an alias or not.